* commit 'c975437ef3338ba38b76b1a07594ccfd6e089709':
  AArch64: Use long for pointers in MTP classes

+++ media/java/android/mtp/MtpDevice.java
@@ -235,7 +235,7 @@ public final class MtpDevice {
     }
 
     // used by the JNI code
-    private int mNativeContext;
+    private long mNativeContext;
 
     private native boolean native_open(String deviceName, int fd);
     private native void native_close();
